Added new run animation texture
I added the new running texture to the Sprite Animation folder under
Assets/Art Assets. Included are the text files (essentially a json???
file which tells the computer to remap the texture to a different
origin, therefore making it  appear animated). Idk how to get it working
fully in unity, but I think you just need to make a flat 3d plane, apply
the texture, and somehow tell it to remap it according to the text
file's code.